PathHelix: Fixes for clearance, non-aligned centers
The safe Z value was calculated with respect to the current hole, but
the tool should better be always retracted to the safe Z value with
respect to the first hole of a series of holes.

Another fix concerns the detection of holes-within-holes, previously it
was assumed that there the centers must align perfectly, this seems to
be the case only up to some numerical precision. Fixed by just allowing
any hole that is wholly contained in the first hole, even it is not
centered.