Sketcher: fix false detection of doubleclick
Double-click detector code was comparing double-click threshold (value =
5) with distance between clicked objects in model space (i.e.
double-click distance threshold was equal to 5 mm in model space
regardless of zoom). The fix changes the comparison to pixels on screen
instead.

This fixes the inability to select sketch entities in quick succession,
when the sketch is small (i.e. less than 40 mm or so).