Crowdin: Experiment to remove obsolete translations from .ts file
Crowdin currently has an issue where typo strings that have been fixed in the FC source haven't been permeating in to the Crowdin UI. This maybe do to discrepency between lupdate and crowdin.  This PR experiments with removing a few lingering obsolete translations in the main .ts file of the Arch and Path WBs. If this works then a follow-up PR will be made to remove all the rest of the obsolete strings.